+++ pyssa/utils_cython.pxd
@@ -1,1 +1,1 @@
-cdef roulette_selection(double[:] prop_list, long[:] Xt)
+cdef roulette_selection(double [:] prop_list, long long [:] Xt)
